Switching to using the notices API instead of the snackbar.
