--- client/tos/snackbar/style.scss
@@ -1,5 +1,0 @@
-.woocommerce-payments__tos-snackbar {
-	position: fixed;
-	bottom: 0;
-	margin: 0 0 20px 20px;
-}
